requirements.txt
absl-py
annotated-types
anyio
asgiref
astunparse
attrs
CacheControl
cachetools
certifi
cffi
charset-normalizer
click
cloudinary
colorama
contourpy
crispy-bootstrap4
cryptography
cycler
django
django-crispy-forms
django-multiselectfield
django-rest-framework-nested
djangorestframework
dm-tree
et-xmlfile
fastapi
flatbuffers
fonttools
gast
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
h11
h5py
httplib2
idna
ImageHash
Jinja2
joblib
keras
kiwisolver
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
missingno
ml-dtypes
msgpack
multimethod
namex
networkx
numpy
openpyxl
opt-einsum
packaging
pandas
Pillow
proto-plus
protobuf
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
PyJWT
pyparsing
python-dateutil
python-dotenv
python-slugify
pytz
PyWavelets
PyYAML
requests
rich
rsa
scipy
seaborn
setuptools
six
sniffio
sqlparse
starlette
tensorflow==2.16.1
ip2geotools
geopy
python-decouple
tangled-up-in-unicode
termcolor
text-unidecode
tqdm
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
visions
werkzeug
wheel
wrapt
phonenumbers
pycryptodome
pyrebase4
mysqlclient